Refactoring to remove dependency of WebglGraphicsDevice on shader chunks
