+++ src/graphics/program-lib/chunks/chunks.js
@@ -133,8 +133,6 @@ import particle_softVS from './particle/vert/particle_soft.js';
 import particle_stretchVS from './particle/vert/particle_stretch.js';
 import particle_TBNVS from './particle/vert/particle_TBN.js';
 import particle_wrapVS from './particle/vert/particle_wrap.js';
-import precisionTestPS from './common/frag/precisionTest.js';
-import precisionTest2PS from './common/frag/precisionTest2.js';
 import reflDirPS from './lit/frag/reflDir.js';
 import reflDirAnisoPS from './lit/frag/reflDirAniso.js';
 import reflectionCCPS from './lit/frag/reflectionCC.js';
@@ -340,8 +338,6 @@ const shaderChunks = {
     particle_stretchVS,
     particle_TBNVS,
     particle_wrapVS,
-    precisionTestPS,
-    precisionTest2PS,
     reflDirPS,
     reflDirAnisoPS,
     reflectionCCPS,
