--- src/graphics/program-lib/chunks/common/frag/precisionTest.js
@@ -1,5 +1,0 @@
-export default /* glsl */`
-void main(void) {
-    gl_FragColor = vec4(2147483648.0);
-}
-`;
